updated options
